style: featured labs layout
