+++ apps/commudle-admin/src/app/app-shared-components/labs-featured/labs-featured.component.scss
