+++ apps/commudle-admin/src/app/feature-modules/homepage/components/homepage-dashboard/homepage-dashboard.component.html
@@ -22,6 +22,9 @@
         <div class="featured-projects">
           <commudle-featured-projects [showCardsHorizontal]="true" [showTag]="false"></commudle-featured-projects>
         </div>
+        <div class="labs-featured">
+          <commudle-labs-featured [showCardsHorizontal]="true"></commudle-labs-featured>
+        </div>
       </div>
     </div>
   </nb-card-body>
